+++ src/Components/Portfolio/PortfolioSection.tsx
@@ -1,14 +1,20 @@
-import React, { useState, useCallback, FC } from "react";
+import React, { useMemo, useEffect, useRef, useState, useCallback, FC } from "react";
 import {
   // WidthProvider,
-  Responsive } from "react-grid-layout";
+  Responsive
+} from "react-grid-layout";
 import WidthProvider from '../../HoC/WidthProvider';
 import styled from "@emotion/styled";
+
+// Import Hooks
+import useDimsReporter from '../../Hooks/useExposeStatelessDimsReporter';
+import useGridLayoutHandlers from '../../Hooks/useGridLayoutHandlers';
+import { useMemoizedResizeSuggestionAlgorithm } from '../../Hooks/useSuggestResize';
+import { useContentDimsAggregators } from '../../Hooks/useContentDimsAggregators';
 
 import PortfolioItemCard from "./PortfolioItem";
 import { withDefaultProps } from "../hoc";
 import useLayoutsState from '../../Hooks/useLayoutsState';
-import useGridLayoutHandlers from '../../Hooks/useGridLayoutHandlers';
 import Typography from '../Typography';
 
 import PortfolioItemInterface from "../../PortfolioItemInterface";
@@ -107,7 +113,7 @@ const ResetLayoutButton = styled(ResetLayoutButtonWithTypography) <ResetLayoutBu
 // WITH WIDTH from WINDOW
 const ResponsiveReactGridLayout = WidthProvider(Responsive);
 
-type RenderProps = (data: PortfolioItemInterface, theme: AppPortfolioItemProps['theme']) => React.ReactNode;
+type RenderProps = (data: PortfolioItemInterface, theme: AppPortfolioItemProps['theme'], refs?: React.RefObject<HTMLElement>[]) => React.ReactNode;
 interface ResponsiveLocalStorageLayoutProps {
   id: string;
   data: PortfolioItemInterface[];
@@ -153,20 +159,62 @@ interface ResponsiveLocalStorageLayoutProps {
 
 // DESIGNER's ENTRYPOINT
 
+// helper function
+// const defaultsMaker = (): {
+//   className: string,
+//   cols: {
+//     lg: number;
+//     md: number;
+//     sm: number;
+//     xs: number;
+//     xxs: number;
+//   },
+//   rowHeight: number,
+//   renderProps: RenderProps,
+//   element_to_render: typeof PortfolioItemCard,
+//  } => {
+//   return { a: 1 };
+// }
+
+interface AppPortfolioSectionDefaults {
+  className: string,
+  cols: {
+    lg: number;
+    md: number;
+    sm: number;
+    xs: number;
+    xxs: number;
+  },
+  rowHeight: number,
+  renderProps: RenderProps,
+  element_to_render: typeof PortfolioItemCard,
+}
+
+type DefaultsType<T> = T extends Partial<ResponsiveLocalStorageLayoutProps> ? T : never;
+// T should be checked that is a Partial<ResponsiveLocalStorageLayoutProps>
+// type DefaultMaker<T> = () 
+
 // used to set up default values for the properties of the component
 // if the client does not provide them
 /** Default Design of breakpoints via 'cols' */
-const defaultProps: Partial<ResponsiveLocalStorageLayoutProps> = {
+const defaultProps: DefaultsType<AppPortfolioSectionDefaults> = {
   className: "layout",
   cols: { lg: 12, md: 10, sm: 6, xs: 4, xxs: 2 },
   // TODO make this dynamic based on the content of the PortfolioItem
   rowHeight: 41, // governs the length each Portfolio Card will cover on the y axis, on initial render,
   // Responsible for Content of Grid Item
-  renderProps: (data, theme) => <AppPortfolioItem data={data} theme={theme} />,
+  renderProps: (data, theme, refs) => <AppPortfolioItem data={data} theme={theme} refs={refs} />,
   // Direct Child Component of LayoutItem that receives the above Render Props Callback
   // Responsible for Styles of Grid Item
-  element_to_render: PortfolioItemCard,  // 2 DIVS
+  element_to_render: PortfolioItemCard,  // 1 DIV, with children the renderProps(data, theme) result
 };
+
+/// Local Type Checking
+// Dimensions Reporter Interface
+type DimsReporter = () => { width: number; height: number };
+type Reducer<S, T> = (acc: S, _: T, index: number) => S;
+type ContentItem = { ref: React.RefObject<HTMLElement>, dimsReporter: DimsReporter };
+type ContentRegistry = Record<string, ContentItem[]>;
 
 // COMPONENT - Reactive Grid Layout (with Local Storage support)
 /**
@@ -186,6 +234,8 @@ const ResponsiveLocalStorageLayout: FC<ResponsiveLocalStorageLayoutProps> = ({
 
   // Code for implementing Saving and Loading Layouts from Local Storage
   const [layouts, setLayouts, saveToLS] = useLayoutsState();
+  // hook for delegating onResize Handler logic to
+  const [resizeAlgorithmCallback] = useMemoizedResizeSuggestionAlgorithm();
 
   // EVENT HANDLERS - RESET BUTTON
   /**
@@ -196,13 +246,82 @@ const ResponsiveLocalStorageLayout: FC<ResponsiveLocalStorageLayoutProps> = ({
   const handleResetLayout = () => {
     setLayouts({});
   };
-  // EVENT HANDLERS - GRID LAYOUT
-  const [handleLayoutChange, handleItemResize] = useGridLayoutHandlers({
+  // EVENT HANDLERS - any GRID LAYOUT CHANGE
+  const [handleLayoutChange] = useGridLayoutHandlers({
+    // Callables (aka functions/callbacks) that handleLayoutChange calls
     setLayouts,
     saveToLS: useCallback((allLayouts: LayoutsObject) => {
       saveToLS("layouts", allLayouts);
     }, [saveToLS]),
   });
+
+  // Intialize Ref storing the ContentRegistry of each Portfolio Item
+  const reducer: Reducer<ContentRegistry, PortfolioItemInterface> = (acc, _, index) => {
+    // we support binding 3 DOM Elements of Portfolio Item Content Elements
+    acc[index.toString()] = Array.from({ length: 3 }, () => {
+      const [ref, dimsReporter] = useDimsReporter();
+      return {
+        ref,
+        dimsReporter,
+      }
+    });
+    return acc;
+  }
+  // DECLARE ContentRegistry Ref, as mapping of Grid Item IDs to Array of 3 ContentItem
+  const contentRegistry = useRef<ContentRegistry>(
+    data.reduce(reducer, {})
+  );
+
+  // create Callbacks to sum Heights and Widths of Portfolio Item Contents
+  const [sumContentHeight, sumContentWidth] = useContentDimsAggregators(contentRegistry.current,
+    'height', 'width'
+  );
+
+  // ON RESIZE EVENT HANDLER
+
+  // Constants
+  const UNIT_LENGTH: number = 50;  // px
+
+  // onResize events fire continuously while User's drag-n-drop on the bottom-right corner of a Grid Item
+  const handleItemResize = (
+    layout,
+    oldItem,
+    newItem,
+    placeholder,
+    event,
+    element,
+  ) => {
+    const index: string = newItem.i.toString();
+
+    // Get potential Height Unit Suggestion, with Content-Aware Adjustment
+    const suggestedUnitValues = resizeAlgorithmCallback(
+      newItem,  // information about Resized Item Dim Units; ie 1, 2, 3, 4, 5, 6
+      {
+        // fit Units to contentHeight, as sum of heights requested from browser-api
+        contentHeight: contentRegistry.current[index].reduce((acc, { dimsReporter }) => acc + dimsReporter().height, 0),
+        // contentWidth: Math.max(...contentRegistry.current[index].map(({ dimsReporter }) => dimsReporter().width)),
+        unit_length: UNIT_LENGTH,
+        width_unit_length: 97.5,
+        // account for Description Component Margin
+        contentAdjustmentOffsetHeight: theme.item.theme.projectDescription.margin * 2,
+        // contentAdjustmentOffsetWidth: -40, // px
+      }
+    );
+    // UPDATE: increase suggested in Height Units
+    if (suggestedUnitValues !== undefined) {
+      if (suggestedUnitValues.unitsHeight) {
+        newItem.h = suggestedUnitValues.unitsHeight;
+        placeholder.h = suggestedUnitValues.unitsHeight;
+      }
+      if (suggestedUnitValues.unitsWidth) {
+        newItem.w = suggestedUnitValues.unitsWidth;
+        placeholder.w = suggestedUnitValues.unitsWidth;
+      }
+    } else {  // DEBUG CODE ELSE
+      console.warn('Content Height is not available, because refs are not attached to the DOM element');
+    }
+
+  };
 
   // CONSTANT: starting width of each Portfolio Item
   const startingWidth = 4;
@@ -229,10 +348,8 @@ const ResponsiveLocalStorageLayout: FC<ResponsiveLocalStorageLayoutProps> = ({
         // we store the Layouts (breakpoints -> layouts) in the component's state, and local storage when user changes the layout (when an item is droped (after a drag-n-drop), when the user resizes an item after they release of the click button), or when the "Reset Layout" button is clicked
         onLayoutChange={handleLayoutChange}
 
-        // handle resize, happening when user drags from bottom right
-        // this runs after user clicks bottom-right and while they keep the mouse clicked. It runs 'continuously' on every pixel moved sort-a-thing.
-        // currently we use simple heuristic to increase height when width decreases
-        onResize={handleItemResize}  // if W <= 2 add 2 H else if W <= 3 add 1 H
+        // handle resize events, which fire 'continuously' while user holds mouse pressed, after clicking on the bottom-right of any Grid Item
+        onResize={handleItemResize}  // we use content-aware algorithm to increase height if inner content requires it
 
         // if true the items will appear and animate/"transition" to match the state Layouts Object
         // if false, the items will appear instantly, without any animation
@@ -260,7 +377,6 @@ const ResponsiveLocalStorageLayout: FC<ResponsiveLocalStorageLayoutProps> = ({
 
           // Initialize the zIndex state for this item
           const [zIndex, setZIndex] = useState(0);
-
           const setStartingHeigth = () => {
             // and a heuristic including the startingWidth and the number of characters in data.description
             // to determine the height of the Portfolio Item
@@ -280,13 +396,14 @@ const ResponsiveLocalStorageLayout: FC<ResponsiveLocalStorageLayoutProps> = ({
             }
             return height
           };
+
           // MEMOIZATION OF GRID ITEMS
           //  - avoids all item re-renders when moving or resizing!
-          //  - zIndex changes only trigger re-render of relevant Item
+          //  - zIndex changes only trigger re-render of its own parent LayoutItem
           const child = React.useMemo(() => {
             return (
               <LayoutItem  // DIV
-                key={index}
+                key={index.toString()}
                 data-grid={{
                   i: index.toString(),
                   w: startingWidth,
@@ -313,10 +430,16 @@ const ResponsiveLocalStorageLayout: FC<ResponsiveLocalStorageLayoutProps> = ({
                 <ZIndexContext.Provider value={{
                   setZIndex
                 }}>
-                  <ResponsiveLayoutItemContent // 2 DIVs
+                  <ResponsiveLayoutItemContent // 1 DIV
                     data={item}
                     // Renders Fragment of Elements with Grid Item Contents
-                    renderProps={(d: PortfolioItemInterface) => { return inputRenderProps(d, theme.item.theme) }}
+                    renderProps={(d: PortfolioItemInterface) => {
+                      return inputRenderProps(
+                        d,
+                        theme.item.theme,
+                        contentRegistry.current[index.toString()].map(({ ref }) => ref as React.RefObject<HTMLElement>),
+                      )
+                    }}
                   />
                 </ZIndexContext.Provider>
               </LayoutItem>